- style fixes - comments added - E2E_P11Check: return value fixed

+++ e2e_utils.c
@@ -14,11 +14,11 @@ uint8_t update_crc8_sae_j1850(unsigned char inpcrc, const uint8_t data)
 
     crc = inpcrc;
     crc ^= data;
-    for (bit = 0; bit < 8; bit++)
+    for (bit = 0u; bit < 8; bit++)
     {
-        if ((crc & 0x80) != 0)
+        if ((crc & 0x80) != 0u)
         {
-            crc <<= 1;
+            crc <<= 1u;
             crc ^= 0x1D;
         }
         else
@@ -27,7 +27,7 @@ uint8_t update_crc8_sae_j1850(unsigned char inpcrc, const uint8_t data)
         }
     }
 
-    return (~crc) & 0xFF;
+    return (uint8_t) (~crc) & 0xFF;
 }
 
 uint8_t crc8_sae_j1850(const uint8_t *data, uint8_t length)
@@ -36,19 +36,19 @@ uint8_t crc8_sae_j1850(const uint8_t *data, uint8_t length)
     uint8_t i, bit;
 
     crc = 0xFF;
-    for (i = 0; i < length; i++)
+    for (i = 0u; i < length; i++)
     {
         crc ^= data[i];
-        for (bit = 0; bit < 8; bit++)
+        for (bit = 0u; bit < 8; bit++)
         {
-            if ((crc & 0x80) != 0)
+            if ((crc & 0x80) != 0u)
             {
-                crc <<= 1;
+                crc <<= 1u;
                 crc ^= 0x1D;
             }
             else
             {
-                crc <<= 1;
+                crc <<= 1u;
             }
         }
     }
